+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Adds a duplicate key checker to the YAML `Loader` that raises an error when a duplicate key is
   found, and points to the file and line number that caused the error. The YAML `Loader` modification
   maintains compliance with the existing JSON validation protocols.
+- Add a simple nuclear plant performance and cost model
 
 ## 0.6 [February 10, 2026]